fix clone3 calling assembly code
Since we were passing values using registers, the could have been used
both as an input and as a clobbering register (since we aren't
specifying the registers that are being used as input). There were cases
when the same register that was being as an input was being overwritten
by another instruction before we could actually use the input. This was
breaking the sycall.

So this is why we have decided to pass all the values using memory
instead.